fix the bug that utilization bound is wrong when scaling down is executed
root cause: the recommended parallelism is the max recommended parallelism in the past scale down interval window(is not the latest parallelism), but we check the utilization bound for the latest metric. Sometimes, the recommended parallelism is inside the utilization bound, but latest utilization bound is outside of utilization bound, and it causes the scale down is executed. It's unexpected.